try to fix tests failing because cleanup not triggered at the end of the testsuite
