+++ tests/src/test/resources/META-INF/services/org.junit.platform.launcher.TestExecutionListener
@@ -1,1 +1,1 @@
-io.apicurio.tests.RegistryDeploymentManager
+io.apicurio.tests.ApicurioRegistryTestExecutionListener
